--- src/pages/interfaces.ts
@@ -1,2 +1,0 @@
-export type StatusType = "error" | "success";
-export type TelegramStatusType = "error" | "info" | "success";
